+++ src/SmartOffice/Controllers/AccountController.cs
@@ -1,0 +1,58 @@
+﻿// -----------------------------------------------------------------------
+// <copyright file="AccountController.cs" company="Microsoft">
+//     Copyright (c) Microsoft Corporation. All rights reserved.
+// </copyright>
+// -----------------------------------------------------------------------
+
+namespace Microsoft.Partner.SmartOffice.Controllers
+{
+    using AspNetCore.Mvc;
+    using Microsoft.AspNetCore.Authentication;
+    using Microsoft.AspNetCore.Authentication.Cookies;
+    using Microsoft.AspNetCore.Authentication.OpenIdConnect;
+
+    public class AccountController : Controller
+    {
+        //
+        // GET: /Account/SignIn
+        [HttpGet]
+        public IActionResult SignIn()
+        {
+            return Challenge(
+                new AuthenticationProperties { RedirectUri = "/" }, OpenIdConnectDefaults.AuthenticationScheme);
+        }
+
+        //
+        // GET: /Account/SignOut
+        [HttpGet]
+        public IActionResult SignOut()
+        {
+            string callbackUrl = Url.Action(nameof(SignedOut), "Account", values: null, protocol: Request.Scheme);
+
+            return SignOut(new AuthenticationProperties { RedirectUri = callbackUrl },
+                CookieAuthenticationDefaults.AuthenticationScheme, OpenIdConnectDefaults.AuthenticationScheme);
+        }
+
+        //
+        // GET: /Account/SignedOut
+        [HttpGet]
+        public IActionResult SignedOut()
+        {
+            if (HttpContext.User.Identity.IsAuthenticated)
+            {
+                // Redirect to home page if the user is authenticated.
+                return RedirectToAction(nameof(HomeController.Index), "Home");
+            }
+
+            return View();
+        }
+
+        //
+        // GET: /Account/AccessDenied
+        [HttpGet]
+        public IActionResult AccessDenied()
+        {
+            return View();
+        }
+    }
+}